Adding Main and 2 base classes
Class FSM has 2 variables input string to contain the given text and a hashmap for the given nodes. There are 2 functions 1 constructor and createNodes to create the amount of nodes needed for the FSM.

Class Node has 2 variables one for the name of the node and a HashMap to contain the connections between the nodes.